+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Release Notes
 
-## [v4.11.0](https://github.com/Sleon4/Lion-Files/compare/v4.10.0...v4.11.0) (2023-06-27)
+## [v5.0.0](https://github.com/lion-packages/files/compare/v4.11.0...v5.0.0) (2023-11-13)
+
+### Added
+- Docker configuration added
+- testing configuration added
+- added test for LionFiles\Zip class
+- added test for LionFiles\Store class
+
+## [v4.11.0](https://github.com/lion-packages/files/compare/v4.10.0...v4.11.0) (2023-06-27)
 
 ### Added
 - added decompress function for LionFiles\Zip class
@@ -8,7 +16,7 @@
 ### Refactoring
 - function responses have been modified
 
-## [v4.10.0](https://github.com/Sleon4/Lion-Files/compare/v4.9.0...v4.10.0) (2023-05-30)
+## [v4.10.0](https://github.com/lion-packages/files/compare/v4.9.0...v4.10.0) (2023-05-30)
 
 ### Added
 - added get function to get files through file_get_contents